Repo init with basic reporter
Starts the repo and introduces the DocutestReporter
which generates a markdown file for each test file
tested via Vitest.

The markdown file leverages Vitest's TaskMeta
adding the "title" and "description" fields
that can be used to add rich content to the markdown file.

Includes a `runVitest` util coming from the Vitest repo.
Unfortunately the code is copied and hardcoded because it
can't be installed in package.json:
https://github.com/vitest-dev/vitest/blob/8b4a44ad16461612fa9884dbe7b831a2f80eacfc/test/test-utils/index.ts